# Conflicts:
#	PITA_Rewards2/app/src/main/java/com/example/pita_rewards2/checkoutActivities/BasketActivity.kt
#	PITA_Rewards2/app/src/main/java/com/example/pita_rewards2/checkoutActivities/CheckoutActivity.kt
#	PITA_Rewards2/app/src/main/java/com/example/pita_rewards2/checkoutActivities/EmployeeActivity.kt
#	PITA_Rewards2/app/src/main/java/com/example/pita_rewards2/mainActivities/Drink_Menu.kt
#	PITA_Rewards2/app/src/main/java/com/example/pita_rewards2/mainActivities/MainActivity.kt

+++ PITA_Rewards2/app/src/main/java/com/example/pita_rewards2/userActivities/UserData.kt
@@ -6,7 +6,7 @@ data class UserData(
     val password: String? = null,
     val phone: String? = null,
     val studentID: String? = null,
-    var points: Int = 0, // mutable int
+    var points: Double = 0.0, // mutable double
     val firstName: String? = null,
     val lastName: String? = null
 )
